refactor: Upgrade codebase for `ecphp/cas-lib` upcoming version 2.
