+++ spec/EcPhp/Ecas/EcasSpec.php
@@ -13,7 +13,11 @@ namespace spec\EcPhp\Ecas;
 
 use EcPhp\CasLib\Cas;
 use EcPhp\CasLib\Response\CasResponseBuilder;
-use EcPhp\Ecas\Response\EcasResponseBuilder;
+use EcPhp\CasLib\Response\Factory\AuthenticationFailureFactory;
+use EcPhp\CasLib\Response\Factory\ProxyFactory;
+use EcPhp\CasLib\Response\Factory\ProxyFailureFactory;
+use EcPhp\CasLib\Response\Factory\ServiceValidateFactory as FactoryServiceValidateFactory;
+use EcPhp\Ecas\Response\Factory\ServiceValidateFactory;
 use Ergebnis\Http\Method;
 use Exception;
 use loophp\psr17\Psr17;
@@ -185,9 +189,11 @@ class EcasSpec extends ObjectBehavior
             new Psr18Client(CasHelper::getHttpClientMock()),
             $psr17,
             new ArrayAdapter(),
-            new EcasResponseBuilder(
-                new CasResponseBuilder(),
-                $psr17
+            new CasResponseBuilder(
+                new AuthenticationFailureFactory(),
+                new ProxyFactory(),
+                new ProxyFailureFactory(),
+                new ServiceValidateFactory(new FactoryServiceValidateFactory(), $psr17)
             )
         );
